pyproject.toml
[project]
name = "ludic-ai"
version = "0.1.0"
description = "A domain-agnostic framework for competence-controlled evaluation of persona-conditioned suboptimization."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scipy>=1.10",
  "statsmodels>=0.14",
  "matplotlib>=3.8",
  "PyYAML>=6.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ludic_ai"]
